core/requirements.txt
echo "Django>=4.0
djangorestframework
djangorestframework-simplejwt" > requirements.txt
